rag_backend/requirements.txt
fastapi
uvicorn
pandas
chromadb
sentence-transformers
python-dotenv
mistralai
httpx
pydantic
